demo-call: Remove unnecessary g_timer_stop()
Calling g_timer_stop() will just set the end time and set the timer inactive.
Instead we can just clear the timer and don't need to worry about the state of
the timer at all.

Also it makes accept/hangup case more symmetric in that accept creates the timer
and hangup destroys it.