display: Always show status text
Do not ever hide it so that it can always be used, no matter what state
the call is in.

Also move the status text label into the box one level higher in the
widget hierarchy which let's us remove a GtkBox because now it only has
a single child.

The XML was not reindented to make review easier.