demo-call: Document time tracking should only stop on disconnect
We want to have the time keep ticking when e.g. on hold.

While functionality wise the demo would never run into any issues,
we should still document intended usage of the timer and that there is a
need for guarding against starting the timer multiple times.

As the demo could be considered the "reference implementation" for how
to do the timer it's usage should be documented.