demo: Remove separator from AdwLeaflet
Not sure exactly what the issue is, but without this change there's a
problem:
1. use a small window width such that the leaflet folds
2. click on any of the sidebar pages
3. click on the `<` button to return to the page list
4. the page list is invisible and the separator expands to fill the
   entire screen

I've removed it in this commit. It comes at the cost of not having a
separator in the header. There is still a separator between the two
content areas.

This seems like a reasonable tradeoff to me. `AdwLeaflet` is deprecated
anyways and will have to be completely replaced later; see
https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/migrating-to-breakpoints.html#replace-adwleaflet